CodeEditorView: add support for navigating the completion window up/down and inserting completions.
CompletionsDisplayView: add support for different tapping gestures, a
single tap selects an item, a double tap picks the suggested
completion.  A tap on the special icon [return] inserts the
completion.

EditedItem: when a completion is triggerd to run, monitor the location
of the selection every 300ms, to ensure that if the user moves away
via other keyboard combinations (left/right arrow, but also assorted
Emacs key combinations handled natively) we cancel the completion.

PathBrowser: make the prefix a little less bold.